Add method_exists check to load data. #1528
Adds a method_exists check to load data dynamically, based on the plural name of the object, eg ‘affiliates’. The check then calls, via call_user_func(), a method named $this->plural . _data. Failing that, the method table_data() is called.

The table_data() method will likely be removed in a later commit with a more helpful error, or at least an accompanying debug log notice.

#1528